requirements.txt
Django==1.11.2
djangorestframework==3.6.3
gunicorn==19.6.0
whitenoise
psycopg2==2.7.2
dj-database-url==0.3.0
dj-static==0.0.6
python-decouple==3